### 2023-05-12 #111 @monocongo
Please review climate_indices (drought monitoring, SPI, SPEI, etc.)
Submitting Author: James Adams (@monocongo)
Package Name: climate_indices
One-Line Description of Package: A suite of climate indices developed for drought.gov at NOAA, useful for drought monitoring.
Repository Link:  https://github.com/monocongo/climate_indices
Version submitted:   Please use the master branch or latest from PyPI
Version accepted: TBD

---

- [ X ] I agree to abide by [pyOpenSci's Code of Conduct][PyOpenSciCodeOfConduct] during the review process and in maintaining my package after should it be accepted.
- [ X ] I have read and will commit to package maintenance after the review as per the [pyOpenSci Policies Guidelines][Commitment].

This package computes commonly used drought indices such as SPI, SPEI, and PET. The individual indices can be computed via an API where precipitation and temperature values are passed in as numpy arrays, and there is a command line option that allows the usage of NetCDF files as input/output format. These were developed while I worked as part of the drought.gov team (NIDIS/NOAA).

	- [ X] Data processing/munging

	- [ X] Geospatial

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

Audience: Climate scientists, hydrologists
Application: Climate monitoring and analysis

climx
Ours seems to be more performant and focused on fewer but more complex drought indices, and in the end, we should merge so as to not duplicate/waste efforts, I just haven't had the time to fold this into their codebase and guide our users in their direction. From what I can tell our package gets more use in the community, but I am out of that loop on that these days. My main concern is that it follows "best practices", and I can mostly handle the software side of that, but I am not a scientist and need help validating that this code is solid from a scientific perspective. Also, the code here was the code used to produce the various monthly indices provided by NOAA, and my initial goal was to make that as transparent as possible via an open-source approach.

- [x] has a continuous integration setup, such as GitHub Actions CircleCI, and/or others.
