### 2023-05-16 #111 @NickleDave
Welcome to pyOpenSci @monocongo and thank you for your submission!

Good news: `climate_indices` passes most of the initial checks, and we think we have already found an editor for this review.

There's just a couple things missing I need you to add:
- [ ] I don't find API documentation -- if that's correct, can you please add it? (e.g. with [autodoc](https://cjolowicz.github.io/posts/hypermodern-python-05-documentation/#generating-api-documentation-with-autodoc))
- [ ] Can you please add install instructions to the README? Just a `pip install climate-indices` is fine

Please let me know when you've added those and we can proceed with the review. Appreciate it!

  - [x] The package imports properly into a standard Python environment `import package-name`.
    - importing finishes without an error but nothing was imported; `climate_indices/__init__.py` does not import anything from the package
    - I tested that the script `process_climate_indices` got installed; it gave me the argparse options as expected, so this suggests that installation does work
  - [ ] API documentation (documentation for your code's functions, classes, methods and attributes): this includes clearly written docstrings with variables defined using a standard docstring format. *We suggest using the [Numpy](https://numpydoc.readthedocs.io/en/latest/format.html#docstring-standard) docstring format*.
  - [x] **Code of Conduct** The package has a `Code of Conduct` file.
- [x] **Automated tests** Package has a testing suite and is tested via GitHub actions or another Continuous Integration service.

---
---
