### 2023-05-22 #111 @monocongo
Thank you for this initial review @NickleDave . I will further document the code with docstrings and generate API docs via `autodoc`.
