### 2023-06-14 #111 @NickleDave
Hi @monocongo just want to check back where you're at with API docs and adding install instructions to the README.

If it's okay with you, I'll open an issue on the climate_indices repo with these to-do items that references this issue, just so we can track changes made as part of the process, and so we can follow the link here to know where you're at with it

We do have an editor that can start the review, but if you aren't able to get to the API docs soon, I totally understand, believe me. If I open the issue on climate_indices that will also help us know when you're ready
