### 2023-06-16 monocongo/climate_indices#521 @NickleDave
Add API docs and install instructions to README
Hi @monocongo just raised that issue related to the pyOpenSci review as discussed here: https://github.com/pyOpenSci/software-submission/issues/111#issuecomment-1591996074

Before we can start the review we'll need
- [ ] API docs, e.g. generated with autodoc
- [ ] brief install instructions in the readme (can be as simple as `pip install climate_indices`

Just let us know when you have a chance to take care of those and we'll get you that review, thank you!
