+++ src/climate_indices/__init__.py
@@ -1,2 +1,0 @@
-# version of the climate_indices package, should match with setup.py
-__version__ = "1.0.10"
