+++ requirements.txt
@@ -1,60 +1,0 @@
-#
-# This file is autogenerated by pip-compile with python 3.7
-# To update, run:
-#
-#    pip-compile --output-file=requirements.txt pyproject.toml
-#
-cloudpickle==2.2.0
-    # via dask
-dask==2022.2.0
-    # via climate-indices (pyproject.toml)
-fsspec==2022.11.0
-    # via dask
-importlib-metadata==5.1.0
-    # via
-    #   numba
-    #   xarray
-llvmlite==0.39.1
-    # via numba
-locket==1.0.0
-    # via partd
-numba==0.56.4
-    # via climate-indices (pyproject.toml)
-numpy==1.22.0
-    # via
-    #   numba
-    #   pandas
-    #   scipy
-    #   xarray
-packaging==22.0
-    # via dask
-pandas==1.3.5
-    # via xarray
-partd==1.3.0
-    # via dask
-python-dateutil==2.8.2
-    # via pandas
-pytz==2022.6
-    # via pandas
-pyyaml==6.0
-    # via dask
-scipy==1.7.3
-    # via climate-indices (pyproject.toml)
-six==1.16.0
-    # via python-dateutil
-toolz==0.12.0
-    # via
-    #   dask
-    #   partd
-typing-extensions==4.4.0
-    # via
-    #   importlib-metadata
-    #   xarray
-xarray==0.20.2
-    # via climate-indices (pyproject.toml)
-zipp==3.11.0
-    # via importlib-metadata
-
-# The following packages are considered to be unsafe in a requirements file:
-# setuptools
-setuptools>=65.5.1 # not directly required, pinned by Snyk to avoid a vulnerability
