+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py37,py38,py39,py310
+envlist = py38,py39,py310,py311
 
 # Define the minimal tox version required to run; if the host
 # tox is less than this the tool with create an environment and
@@ -18,8 +18,9 @@ setenv = NUMBA_DISABLE_JIT = 1
 deps =
     codecov
     coverage
+    poetry
     pytest
 commands =
-    python setup.py check -m -s
+    poetry install
     coverage run -m pytest tests {posargs}
     codecov
