+++ environment.yml
@@ -1,51 +1,0 @@
-name: climate_indices
-channels:
-  - conda-forge
-  - defaults
-dependencies:
-  - _libgcc_mutex=0.1=conda_forge
-  - _openmp_mutex=4.5=2_gnu
-  - ca-certificates=2022.12.7=ha878542_0
-  - ld_impl_linux-64=2.39=hcc3a1bd_1
-  - libffi=3.4.2=h7f98852_5
-  - libgcc-ng=12.2.0=h65d4601_19
-  - libgomp=12.2.0=h65d4601_19
-  - libnsl=2.0.0=h7f98852_0
-  - libsqlite=3.40.0=h753d276_0
-  - libstdcxx-ng=12.2.0=h46fd767_19
-  - libzlib=1.2.13=h166bdaf_4
-  - ncurses=6.3=h27087fc_1
-  - openssl=3.0.7=h0b41bf4_1
-  - pip=22.3.1=pyhd8ed1ab_0
-  - python=3.7.12=hf930737_100_cpython
-  - readline=8.1.2=h0f457ee_0
-  - setuptools=65.5.1=pyhd8ed1ab_0
-  - sqlite=3.40.0=h4ff8645_0
-  - tk=8.6.12=h27826a3_0
-  - wheel=0.38.4=pyhd8ed1ab_0
-  - xz=5.2.6=h166bdaf_0
-  - pip:
-    - cftime==1.6.2
-    - cloudpickle==2.2.0
-    - dask==2022.2.0
-    - fsspec==2022.11.0
-    - importlib-metadata==5.1.0
-    - llvmlite==0.39.1
-    - locket==1.0.0
-    - nco==1.0.0
-    - netcdf4==1.6.2
-    - numba==0.56.4
-    - numpy==1.21.6
-    - packaging==22.0
-    - pandas==1.3.5
-    - partd==1.3.0
-    - python-dateutil==2.8.2
-    - pytz==2022.6
-    - pyyaml==6.0
-    - scipy==1.7.3
-    - six==1.16.0
-    - toolz==0.12.0
-    - typing-extensions==4.4.0
-    - xarray==0.20.2
-    - zipp==3.11.0
-prefix: /home/james/miniconda3/envs/climate_indices
