### 2023-06-29 3e1f163
[522] new unit tests workflow for running tests against Python versions 3.8 - 3.11
