+++ .github/workflows/unit-tests-workflow.yml
@@ -1,0 +1,26 @@
+name: Python Unit Tests
+
+on: [push, pull_request]
+
+jobs:
+  test:
+    strategy:
+      matrix:
+        python-version: [3.11, 3.10, 3.9, 3.8]
+
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v2
+        with:
+          python-version: ${{ matrix.python-version }}
+
+      - name: Install dependencies
+        run: python -m pip install --upgrade pip poetry
+
+      - name: Set up project
+        run: poetry install
+
+      - name: Run unit tests
+        run: poetry run pytest
