### 2023-06-29 e3e0e48
[522] added quotes to version specs
