### 2023-06-29 70a9260
[522]  now executing poetry via Python as a module
