+++ .github/workflows/unit-tests-workflow.yml
@@ -20,7 +20,7 @@ jobs:
         run: python -m pip install --upgrade pip poetry
 
       - name: Set up project
-        run: poetry install
+        run: python -m poetry install
 
       - name: Run unit tests
-        run: poetry run pytest
+        run: python -m poetry run pytest
