### 2023-06-29 2789a39
[522] added cd $GITHUB_WORKSPACE in order to fix missing pyproject error in GitHub Action
