+++ .github/workflows/unit-tests-workflow.yml
@@ -20,7 +20,9 @@ jobs:
         run: python -m pip install --upgrade pip poetry
 
       - name: Set up project
-        run: python -m poetry install
+        run: |
+          cd $GITHUB_WORKSPACE
+          python -m poetry install
 
       - name: Run unit tests
         run: python -m poetry run pytest
