### 2023-06-29 257c7f2
[522] changed Python setup version from v3 to v2
