+++ .github/workflows/unit-tests-workflow.yml
@@ -20,9 +20,7 @@ jobs:
         run: python -m pip install --upgrade pip poetry
 
       - name: Set up project
-        run: |
-          cd $GITHUB_WORKSPACE
-          poetry install
+        run: poetry install
 
       - name: Run unit tests
         run: poetry run pytest
