### 2023-06-29 77b53bf
[522] added a change directory command to maybe fix the unfound pyproject issue
