+++ .github/workflows/unit-tests-workflow.yml
@@ -6,7 +6,7 @@ jobs:
   test:
     strategy:
       matrix:
-        python-version: ['3.11', '3.10', '3.9', '3.8']
+        python-version: ['3.8', '3.9', '3.10', '3.11']
 
     runs-on: ubuntu-latest
 
@@ -20,7 +20,9 @@ jobs:
         run: python -m pip install --upgrade pip poetry
 
       - name: Set up project
-        run: poetry install
+        run: |
+          cd ..
+          poetry install
 
       - name: Run unit tests
         run: poetry run pytest
