### 2023-06-29 2ad6f69
[522] added a env variable setting to disable Numba JIT for tests
