### 2023-06-29 20578ad
[522] simplified installation instructions, removed developers section (which was mostly redundant)
