### 2023-06-29 56b7760
[522] bumped version to 1.1.0 since so many changes/updates
