+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = 'poetry.core.masonry.api'
 
 [tool.poetry]
 name = 'climate_indices'
-version = '1.0.16'
+version = '1.1.0'
 description = 'Reference implementations of various climate indices typically used for drought monitoring'
 authors = ['James Adams <monocongo@gmail.com>']
 readme = 'README.md'
@@ -33,7 +33,6 @@ xarray = '2023.1.0'
 
 [tool.poetry.dev-dependencies]
 pytest = '*'
-tox = '*'
 
 [tool.pytest.ini_options]
 filterwarnings = [
