### 2023-06-29 cbdd4ff
[522] changed types from numpy floats to plain floats
