### 2023-06-29 344fcff
[522] removed tox and previous GitHub workflow, updated numba and python versions in pyproject.toml
