+++ pyproject.toml
@@ -26,8 +26,8 @@ packages = [{include = 'climate_indices', from = 'src'}]
 
 [tool.poetry.dependencies]
 dask = '2022.2.0'
-numba = '0.56.4'
-python = '>=3.8,<=3.11'
+numba = '0.57.0'
+python = '^3.11'
 scipy = '1.9.3'
 xarray = '2023.1.0'
 
