+++ tox.ini
@@ -1,26 +1,0 @@
-[tox]
-envlist = py38,py39,py310,py311
-
-# Define the minimal tox version required to run; if the host
-# tox is less than this the tool with create an environment and
-# provision it with a tox that satisfies it under provision_tox_env.
-# At least this version is needed for PEP 517/518 support.
-minversion = 3.3.0
-
-# Activate isolated build environment. tox will use a virtual environment
-# to build a source distribution from the source tree. For build tools and
-# arguments use the pyproject.toml file as specified in PEP-517 and PEP-518.
-isolated_build = true
-
-[testenv]
-passenv = CODECOV_TOKEN
-setenv = NUMBA_DISABLE_JIT = 1
-deps =
-    codecov
-    coverage
-    poetry
-    pytest
-commands =
-    poetry install
-    coverage run -m pytest tests {posargs}
-    codecov
