### 2023-06-29 25d967d
[522] added caret constraint operator to the Python version for poetry
