### 2023-06-29 2ddf03d
[522] set operator of the Python version for poetry to be <3.12
