### 2023-06-29 c372726
[522] renamed workflow to tests
