+++ .github/workflows/unit-tests-workflow.yml
@@ -1,4 +1,4 @@
-name: Python Unit Tests
+name: tests
 
 on: [push, pull_request]
 
