+++ README.md
@@ -20,6 +20,7 @@ The following indices are provided:
   or [Hargreaves](http://dx.doi.org/10.13031/2013.26773) equations
 - [PNP](http://www.droughtmanagement.info/percent-of-normal-precipitation/),
   Percentage of Normal Precipitation
+- [PCI](https://www.tandfonline.com/doi/abs/10.1111/J.0033-0124.1980.00300.X), Precipitation Concentration Index
 
 The following are provided as experimental/development versions only, not fully vetted nor suitable for research purposes:
 
