### 2023-06-30 e534ebb
[522] removed numba, cleanups and formatting
