### 2023-06-30 9c92f2f
[522] removed the section describing the Palmer indices from the README
