### 2023-06-30 46e9251
[522] removed Palmer related files and test fixtures
