+++ tests/conftest.py
@@ -84,16 +84,6 @@ def latitude_degrees():
 
 
 @pytest.fixture(scope="module")
-def awc_inches():
-    return _AWC_INCHES
-
-
-@pytest.fixture(scope="module")
-def palmer_awc():
-    return _AWC_INCHES_PALMER
-
-
-@pytest.fixture(scope="module")
 def precips_mm_monthly():
     return np.load(
         os.path.join(os.path.split(__file__)[0],
@@ -265,226 +255,226 @@ def spi_6_month_pearson3():
     )
 
 
-@pytest.fixture(scope="module")
-def palmer_pet():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "palmer_pet.npy")
-    )
+# @pytest.fixture(scope="module")
+# def palmer_pet():
+#     return np.load(
+#         os.path.join(os.path.split(__file__)[0],
+#                      "fixture",
+#                      "palmer_pet.npy")
+#     )
+#
+#
+# @pytest.fixture(scope="module")
+# def palmer_precip():
+#     return np.load(
+#         os.path.join(os.path.split(__file__)[0],
+#                      "fixture",
+#                      "palmer_precip.npy")
+#     )
+#
+#
+# @pytest.fixture(scope="module")
+# def palmer_et():
+#     return np.load(
+#         os.path.join(os.path.split(__file__)[0],
+#                      "fixture",
+#                      "palmer_et.npy")
+#     )
+#
+#
+# @pytest.fixture(scope="module")
+# def palmer_pr():
+#     return np.load(
+#         os.path.join(os.path.split(__file__)[0],
+#                      "fixture",
+#                      "palmer_pr.npy")
+#     )
+#
+#
+# @pytest.fixture(scope="module")
+# def palmer_r():
+#     return np.load(
+#         os.path.join(os.path.split(__file__)[0],
+#                      "fixture",
+#                      "palmer_r.npy")
+#     )
+#
+#
+# @pytest.fixture(scope="module")
+# def palmer_ro():
+#     return np.load(
+#         os.path.join(os.path.split(__file__)[0],
+#                      "fixture",
+#                      "palmer_ro.npy")
+#     )
+#
+#
+# @pytest.fixture(scope="module")
+# def palmer_pro():
+#     return np.load(
+#         os.path.join(os.path.split(__file__)[0],
+#                      "fixture",
+#                      "palmer_pro.npy")
+#     )
+#
+#
+# @pytest.fixture(scope="module")
+# def palmer_l():
+#     return np.load(
+#         os.path.join(os.path.split(__file__)[0],
+#                      "fixture",
+#                      "palmer_l.npy")
+#     )
+#
+#
+# @pytest.fixture(scope="module")
+# def palmer_pl():
+#     return np.load(
+#         os.path.join(os.path.split(__file__)[0],
+#                      "fixture",
+#                      "palmer_pl.npy")
+#     )
+#
+#
+# @pytest.fixture(scope="module")
+# def palmer_pdsi_monthly():
+#     return np.load(
+#         os.path.join(os.path.split(__file__)[0],
+#                      "fixture",
+#                      "palmer_pdsi.npy")
+#     )
+#
+#
+# @pytest.fixture(scope="module")
+# def palmer_pdsi_from_scpdsi_monthly():
+#     return np.load(
+#         os.path.join(
+#             os.path.split(__file__)[0],
+#             "fixture",
+#             "palmer_pdsi_from_scpdsi.npy"
+#         )
+#     )
+#
+#
+# @pytest.fixture(scope="module")
+# def palmer_phdi_monthly():
+#     return np.load(
+#         os.path.join(os.path.split(__file__)[0],
+#                      "fixture",
[…]
