### 2023-06-30 83ae3f6
[522] removed numba, cleanups and formatting
