### 2023-06-30 de72bfa
[522] added CHANGELOG, removed superfluous notebook files
