+++ src/climate_indices/__spi__.py
@@ -29,12 +29,10 @@ _KEY_RESULT = "result_array"
 # global dictionary to contain shared arrays for use by worker processes
 _global_shared_arrays = {}
 
-# ------------------------------------------------------------------------------
 # Retrieve logger and set desired logging level
 _logger = utils.get_logger(__name__, logging.INFO)
 
 
-# ------------------------------------------------------------------------------
 class InputType(Enum):
     """
     Enumeration type for differentiating between gridded, timeseries, and US
@@ -46,7 +44,6 @@ class InputType(Enum):
     timeseries = 3
 
 
-# ------------------------------------------------------------------------------
 def init_worker(arrays_and_shapes):
     """
     Initialization function that assigns named arrays into the global variable.
@@ -62,7 +59,6 @@ def init_worker(arrays_and_shapes):
     _global_shared_arrays = arrays_and_shapes
 
 
-# ------------------------------------------------------------------------------
 def _validate_args(args):
     """
     Validate the processing settings to confirm that proper argument
@@ -202,7 +198,6 @@ def _validate_args(args):
     return input_type
 
 
-# ------------------------------------------------------------------------------
 def _build_arguments(keyword_args):
     """
     Builds a dictionary of function arguments appropriate to the index to be computed.
@@ -224,11 +219,10 @@ def _build_arguments(keyword_args):
     return function_arguments
 
 
-# ------------------------------------------------------------------------------
 def _get_variable_attributes(
-        distribution: indices.Distribution,
-        scale: int,
-        periodicity: compute.Periodicity,
+    distribution: indices.Distribution,
+    scale: int,
+    periodicity: compute.Periodicity,
 ):
 
     attrs = {
@@ -242,12 +236,11 @@ def _get_variable_attributes(
     return attrs
 
 
-# ------------------------------------------------------------------------------
 def _drop_data_into_shared_arrays_grid(
-        dataset_climatology: xr.Dataset,
-        dataset_fitting: xr.Dataset,
-        var_names_climate: List[str],
-        periodicity: compute.Periodicity,
+    dataset_climatology: xr.Dataset,
+    dataset_fitting: xr.Dataset,
+    var_names_climate: List[str],
+    periodicity: compute.Periodicity,
 ):
     """
     Copies arrays from an xarray Dataset into shared memory arrays.
@@ -340,20 +333,6 @@ def _drop_data_into_shared_arrays_grid(
                 _logger.error(message)
                 raise ValueError(message)
 
-        # # convert daily values into 366-day years
-        # if periodicity == compute.Periodicity.daily:
-        #     initial_year = int(dataset_climatology["time"][0].dt.year)
-        #     final_year = int(dataset_climatology["time"][-1].dt.year)
-        #     total_years = final_year - initial_year + 1
-        #     var_values = np.apply_along_axis(utils.transform_to_366day,
-        #                                      len(dims) - 1,
-        #                                      dataset_fitting[var_name].values,
-        #                                      initial_year,
-        #                                      total_years)
-        #
-        # else:  # assumed to be monthly
-        #     var_values = dataset_fitting[var_name].values
-
         var_values = dataset_fitting[var_name].values
 
         # create a shared memory array, wrap it as a numpy array and
@@ -370,7 +349,6 @@ def _drop_data_into_shared_arrays_grid(
         }
 
 
-# ------------------------------------------------------------------------------
 def _drop_data_into_shared_arrays_divisions(
         dataset: xr.Dataset,
[…]
