### 2023-06-30 2c2d2e6
[522] cleanups and formatting, removal of remaining Palmer codes
