+++ src/climate_indices/palmer.py
@@ -1,2358 +1,0 @@
-import collections
-import logging
-import math
-import warnings
-
-import numpy as np
-
-from climate_indices import utils
-
-# Retrieve logger and set desired logging level
-_logger = utils.get_logger(__name__, logging.INFO)
-
-
-_PDSI_MIN = -4.0
-_PDSI_MAX = 4.0
-
-# ignore all warnings
-warnings.simplefilter("ignore", Warning)
-
-
-def _water_balance(awc, pet, precip):
-    """
-    Performs a water balance accounting for a location which accounts
-    for several monthly water balance variables, calculated based on
-    precipitation, potential evapotranspiration, and available water
-    capacity of the soil.
-
-    Input arrays are expected to be the same size, corresponding to
-    the total number of months.
-
-    :param awc: available water capacity (total, including top/surface inch),
-        in inches
-    :param pet: potential evapotranspiration, in inches
-    :param precip: precipitation, in inches
-    :return: seven numpy arrays with values for evapotranspiration,
-        potential recharge, recharge, runoff, potential runoff, loss,
-        and potential loss
-    """
-
-    # flatten timeseries to a 1-D array
-    pet = pet.flatten()
-    precip = precip.flatten()
-
-    total_months = pet.shape[0]
-
-    # allocate arrays for the water balance values
-    ET = np.zeros((total_months,))
-    PR = np.zeros((total_months,))
-    R = np.zeros((total_months,))
-    Rs = np.zeros((total_months,))
-    Ru = np.zeros((total_months,))
-    RO = np.zeros((total_months,))
-    PRO = np.zeros((total_months,))
-    S = np.zeros((total_months,))
-    Ss = np.zeros((total_months,))
-    Su = np.zeros((total_months,))
-    L = np.zeros((total_months,))
-    Ls = np.zeros((total_months,))
-    Lu = np.zeros((total_months,))
-    PL = np.zeros((total_months,))
-    PLs = np.zeros((total_months,))
-    PLu = np.zeros((total_months,))
-
-    # A is the difference between the soil moisture in the surface soil layer and the potential evapotranspiration.
-    A = np.zeros((total_months,))
-
-    # B is the difference between the precipitation and potential evapotranspiration, i.e. the excess precipitation
-    B = np.zeros((total_months,))
-
-    # C is the amount of room (in inches) in the surface soil layer that can be recharged with precipitation
-    C = np.zeros((total_months,))
-
-    # D is the amount of excess precipitation (in inches) that is left over after the surface soil layer is recharged
-    D = np.zeros((total_months,))
-
-    # E is the amount of room (in inches) in the underlying soil layer
-    # that is available to be recharged with excess precipitation
-    E = np.zeros((total_months,))
-
-    # NOTE: SOIL MOISTURE STORAGE IS HANDLED BY DIVIDING THE SOIL INTO TWO
-    # LAYERS AND ASSUMING THAT 1 INCH OF WATER CAN BE STORED IN THE SURFACE
-    # LAYER. awc IS THE COMBINED AVAILABLE MOISTURE CAPACITY IN BOTH SOIL
-    # LAYERS. THE UNDERLYING LAYER HAS AN AVAILABLE CAPACITY THAT DEPENDS
-    # ON THE SOIL CHARACTERISTICS OF THE LOCATION. THE SOIL MOISTURE
-    # STORAGE WITHIN THE SURFACE LAYER (UNDERLYING LAYER) IS THE AMOUNT OF
-    # AVAILABLE MOISTURE STORED AT THE BEGINNING OF THE MONTH IN THE
[…]
