### 2023-06-30 4e6d245
refactor: autofix issues in 2 files
Resolved issues in the following files with DeepSource Autofix:
1. src/climate_indices/compute.py
2. src/climate_indices/lmoments.py
