### 2023-06-30 1b9aa4a
[522] additional docstrings for API generation, cleanups and formatting
