+++ src/climate_indices/__init__.py
@@ -1,0 +1,1 @@
+"""Climate indices for drought monitoring"""
