+++ src/climate_indices/__main__.py
@@ -1,3 +1,5 @@
+"""Command-line interface for climate indices processing"""
+
 import argparse
 from collections import Counter
 from datetime import datetime
@@ -5,6 +7,7 @@ from enum import Enum
 import logging
 import multiprocessing
 import os
+from typing import List
 
 import numpy as np
 import scipy.constants
@@ -28,12 +31,10 @@ _KEY_RESULT_ZINDEX = "result_array_zindex"
 # global dictionary to contain shared arrays for use by worker processes
 _global_shared_arrays = {}
 
-# ------------------------------------------------------------------------------
 # Retrieve logger and set desired logging level
 _logger = utils.get_logger(__name__, logging.INFO)
 
 
-# ------------------------------------------------------------------------------
 class InputType(Enum):
     """
     Enumeration type for differentiating between gridded, timeseriesn and US
@@ -45,7 +46,6 @@ class InputType(Enum):
     timeseries = 3
 
 
-# ------------------------------------------------------------------------------
 def init_worker(arrays_and_shapes):
     """
     Initialization function that assigns named arrays into the global variable.
@@ -61,7 +61,6 @@ def init_worker(arrays_and_shapes):
     _global_shared_arrays = arrays_and_shapes
 
 
-# ------------------------------------------------------------------------------
 def _validate_args(args):
     """
     Validate the processing settings to confirm that proper argument
@@ -419,7 +418,6 @@ def _validate_args(args):
     return input_type
 
 
-# ------------------------------------------------------------------------------
 def _get_scale_increment(args_dict):
 
     if args_dict["periodicity"] == compute.Periodicity.daily:
@@ -432,7 +430,6 @@ def _get_scale_increment(args_dict):
     return scale_increment
 
 
-# ------------------------------------------------------------------------------
 def _log_status(args_dict):
 
     # get the scale increment for use in later log messages
@@ -466,7 +463,6 @@ def _log_status(args_dict):
     return True
 
 
-# ------------------------------------------------------------------------------
 def _build_arguments(keyword_args):
     """
     Builds a dictionary of function arguments appropriate to the index to be computed.
@@ -509,7 +505,6 @@ def _build_arguments(keyword_args):
     return function_arguments
 
 
-# ------------------------------------------------------------------------------
 def _get_variable_attributes(args_dict):
 
     if args_dict["index"] == "spi":
@@ -571,11 +566,12 @@ def _get_variable_attributes(args_dict):
     return var_name, attrs
 
 
-# ------------------------------------------------------------------------------
-def _drop_data_into_shared_arrays_grid(dataset: xr.Dataset,
-                                       var_names: list,
-                                       periodicity: compute.Periodicity,
-                                       data_start_year: int):
+def _drop_data_into_shared_arrays_grid(
+    dataset: xr.Dataset,
+    var_names: list,
+    periodicity: compute.Periodicity,
+    data_start_year: int,
+):
 
     output_shape = None
 
@@ -638,9 +634,10 @@ def _drop_data_into_shared_arrays_grid(dataset: xr.Dataset,
     return output_shape
 
 
-# ------------------------------------------------------------------------------
-def _drop_data_into_shared_arrays_divisions(dataset,
-                                            var_names):
+def _drop_data_into_shared_arrays_divisions(
+    dataset,
+    var_names: List[str],
+):
 
     output_shape = None
 
@@ -684,7 +681,6 @@ def _drop_data_into_shared_arrays_divisions(dataset,
     return output_shape
 
 
-# ------------------------------------------------------------------------------
 def _compute_write_index(keyword_arguments):
     """
     Computes a climate index and writes the result into a corresponding NetCDF.
[…]
