### 2023-07-01 2fa0a86
[522] removed numba and Palmers-related info
