### 2023-07-01 71d7cac
[521] added dependencies for autodoc
