### 2023-07-02 monocongo/climate_indices#521 @monocongo
I have detailed installation instructions available now, and trying to get the API docs finished with autodoc, hopefully in the next day or so. This has led me to consider a revamp of the project docs. Is there an example project that has a good README.md and/or index.rst I can use as a template? Please advise if there's a "model citizen" project in the pyOpenSci community I can emulate.
