+++ docs/reference.rst
@@ -1,0 +1,39 @@
+Reference
+=========
+
+.. contents::
+    :local:
+    :backlinks: none
+
+
+climate_indices.compute
+--------------------------
+
+.. automodule:: climate_indices.compute
+   :members:
+
+
+climate_indices.eto
+----------------------------
+
+.. automodule:: climate_indices.eto
+   :members:
+
+climate_indices.indices
+--------------------------
+
+.. automodule:: climate_indices.indices
+   :members:
+
+
+climate_indices.lmoments
+----------------------------
+
+.. automodule:: climate_indices.lmoments
+   :members:
+
+climate_indices.utils
+----------------------------
+
+.. automodule:: climate_indices.utils
+   :members:
