+++ docs/reference.rst
@@ -1,4 +1,4 @@
-Reference
+API Reference
 =========
 
 .. contents::
