+++ pyproject.toml
@@ -34,6 +34,7 @@ xarray = '2023.1.0'
 
 [tool.poetry.dev-dependencies]
 pytest = '*'
+toml = '0.10.2'
 
 [tool.poetry.group.dev.dependencies]
 sphinx-autodoc-typehints = "^1.23.3"
