### 2023-07-06 monocongo/climate_indices#521 @NickleDave
Hi @monocongo thanks for letting me know.

> Is there an example project that has a good README.md and/or index.rst I can use as a template?

Good question.
All the packages are listed here and their READMEs should give you an idea of what we look for.
https://www.pyopensci.org/python-packages.html
I'm hesitant to say who a "model citizen" would be although I think we could add some recommendations to our guide.
My own opinion, not pyOpenSci's:
- pynteny has a good README: https://github.com/Robaina/Pynteny
- see also devicely: https://github.com/hpi-dhc/devicely
- I like how PyGMT does their docs, notice how they just include README.rst so the two don't go out of sync: https://raw.githubusercontent.com/GenericMappingTools/pygmt/main/doc/index.rst (you could do the same thing with a README.md if you used MyST)

Does that help?
