### 2023-07-10 monocongo/climate_indices#521 @monocongo
Yes, helpful guidance, thanks @NickleDave !
