+++ .github/workflows/deepsource_coverage_report.yml
@@ -1,0 +1,12 @@
+  - name: Report results to DeepSource
+    run: |
+    # Tests have finished and a test coverage report is available
+
+    # Install deepsource CLI
+      curl https://deepsource.io/cli | sh
+
+    # From the root directory, run the report coverage command
+      ./bin/deepsource report --analyzer test-coverage --key python --value-file ./path/to/report
+
+    env:
+      DEEPSOURCE_DSN: ${{ secrets.DEEPSOURCE_DSN }}
