+++ .deepsource.toml
@@ -1,15 +1,0 @@
-version = 1
-
-test_patterns = [
-  "tests/**",
-  "test_*.py"
-]
-
-[[analyzers]]
-name = "python"
-enabled = true
-dependency_file_paths = [
-  "requirements.txt"
-]
-  [analyzers.meta]
-  runtime_version = "3.x.x"
