+++ .github/workflows/unit-tests-workflow.yml
@@ -12,10 +12,10 @@ jobs:
 
     steps:
       - name: Checkout code
-#        uses: actions/checkout@v2
-        uses: actions/checkout@v3
-        with:
-          ref: ${{ github.event.pull_request.head.sha }}
+        uses: actions/checkout@v2
+#        uses: actions/checkout@v3
+#        with:
+#          ref: ${{ github.event.pull_request.head.sha }}
 
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v2
