### 2023-08-04 461dfdd
[526] fixed type hints for a function
