+++ docs/source/modules.rst
@@ -1,0 +1,7 @@
+climate_indices
+===============
+
+.. toctree::
+   :maxdepth: 4
+
+   tests
