+++ docs/source/tests.rst
@@ -1,0 +1,53 @@
+tests package
+=============
+
+Submodules
+----------
+
+tests.conftest module
+---------------------
+
+.. automodule:: tests.conftest
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+tests.test\_compute module
+--------------------------
+
+.. automodule:: tests.test_compute
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+tests.test\_eto module
+----------------------
+
+.. automodule:: tests.test_eto
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+tests.test\_indices module
+--------------------------
+
+.. automodule:: tests.test_indices
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+tests.test\_utils module
+------------------------
+
+.. automodule:: tests.test_utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: tests
+   :members:
+   :undoc-members:
+   :show-inheritance:
