### 2023-08-09 e48ffa6
[526] bumped the workflow checkout version, now using SHA of head commit
