### 2023-08-09 ead6196
[526] refactored pearson_parameters to reduce cognitive complexity
