+++ src/climate_indices/compute.py
@@ -224,6 +224,37 @@ def _probability_of_zero(
     return probabilities_of_zero
 
 
+# +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
+def reshape_values(values, periodicity):
+    if periodicity is Periodicity.monthly:
+        return utils.reshape_to_2d(values, 12)
+    elif periodicity is Periodicity.daily:
+        return utils.reshape_to_2d(values, 366)
+    else:
+        raise ValueError(f"Invalid periodicity argument: {periodicity}")
+
+def validate_values_shape(values):
+    if len(values.shape) != 2 or values.shape[1] not in (12, 366):
+        _log_and_raise_shape_error(shape=values.shape)
+    return values.shape[1]
+
+def adjust_calibration_years(data_start_year, data_end_year, calibration_start_year, calibration_end_year):
+    if (calibration_start_year < data_start_year) or (calibration_end_year > data_end_year):
+        return data_start_year, data_end_year
+    return calibration_start_year, calibration_end_year
+
+def calculate_time_step_params(time_step_values):
+    number_of_zeros, number_of_non_missing = utils.count_zeros_and_non_missings(time_step_values)
+    if (number_of_non_missing - number_of_zeros) < 4:
+        return 0.0, 0.0, 0.0, 0.0
+
+    probability_of_zero = number_of_zeros / number_of_non_missing if number_of_zeros > 0 else 0.0
+
+    if (number_of_non_missing - number_of_zeros) > 3:
+        params = lmoments.fit(time_step_values)
+        return probability_of_zero, params["loc"], params["scale"], params["skew"]
+    return 0.0, 0.0, 0.0, 0.0
+
 def pearson_parameters(
     values: np.ndarray,
     data_start_year: int,
@@ -254,95 +285,149 @@ def pearson_parameters(
         returned array 3 :second Pearson Type III distribution parameter (scale)
         returned array 4: third Pearson Type III distribution parameter (skew)
     """
-
-    # reshape precipitation values to (years, 12) for monthly,
-    # or to (years, 366) for daily
-    if periodicity is Periodicity.monthly:
-
-        values = utils.reshape_to_2d(values, 12)
-
-    elif periodicity is Periodicity.daily:
-
-        values = utils.reshape_to_2d(values, 366)
-
-    else:
-
-        raise ValueError("Invalid periodicity argument: %s" % periodicity)
-
-    # validate that the values array has shape: (years, 12) for monthly or (years, 366) for daily
-    if len(values.shape) != 2:
-        _log_and_raise_shape_error(shape=values.shape)
-
-    else:
-
-        time_steps_per_year = values.shape[1]
-        if time_steps_per_year not in (12, 366):
-            _log_and_raise_shape_error(shape=values.shape)
-
-    # determine the end year of the values array
+    values = reshape_values(values, periodicity)
+    time_steps_per_year = validate_values_shape(values)
     data_end_year = data_start_year + values.shape[0]
-
-    # make sure that we have data within the full calibration period,
-    # otherwise use the full period of record
-    if (calibration_start_year < data_start_year) or \
-            (calibration_end_year > data_end_year):
-        calibration_start_year = data_start_year
-        calibration_end_year = data_end_year
-
-    # get the year axis indices corresponding to
-    # the calibration start and end years
+    calibration_start_year, calibration_end_year = adjust_calibration_years(data_start_year, data_end_year, calibration_start_year, calibration_end_year)
     calibration_begin_index = calibration_start_year - data_start_year
     calibration_end_index = (calibration_end_year - data_start_year) + 1
-
-    # get the values for the current calendar time step
-    # that fall within the calibration years period
[…]
