+++ tests/test_compute.py
@@ -22,102 +22,122 @@ logging.disable(logging.CRITICAL)
     "calibration_year_start_daily",
     "calibration_year_end_daily",
     "transformed_gamma_monthly",
-    "transformed_gamma_daily")
+    "transformed_gamma_daily",
+)
 def test_transform_fitted_gamma(
-        precips_mm_monthly,
-        precips_mm_daily,
-        data_year_start_monthly,
-        data_year_end_monthly,
-        data_year_start_daily,
-        calibration_year_start_monthly,
-        calibration_year_end_monthly,
-        calibration_year_start_daily,
-        calibration_year_end_daily,
-        transformed_gamma_monthly,
-        transformed_gamma_daily):
+    precips_mm_monthly,
+    precips_mm_daily,
+    data_year_start_monthly,
+    data_year_end_monthly,
+    data_year_start_daily,
+    calibration_year_start_monthly,
+    calibration_year_end_monthly,
+    calibration_year_start_daily,
+    calibration_year_end_daily,
+    transformed_gamma_monthly,
+    transformed_gamma_daily,
+):
     """
     Test for the compute.transform_fitted_gamma() function
     """
 
     # confirm that an input array of all NaNs results in the same array returned
     all_nans = np.full(precips_mm_monthly.shape, np.NaN)
-    computed_values = compute.transform_fitted_gamma(all_nans,
-                                                     data_year_start_monthly,
-                                                     data_year_start_monthly,
-                                                     data_year_end_monthly,
-                                                     compute.Periodicity.monthly)
-    np.testing.assert_allclose(computed_values,
-                               all_nans,
-                               equal_nan=True,
-                               err_msg="Gamma fit/transform not handling "
-                                       "all-NaN arrays as expected")
+    computed_values = compute.transform_fitted_gamma(
+        all_nans,
+        data_year_start_monthly,
+        data_year_start_monthly,
+        data_year_end_monthly,
+        compute.Periodicity.monthly,
+    )
+    np.testing.assert_allclose(
+        computed_values,
+        all_nans,
+        equal_nan=True,
+        err_msg="Gamma fit/transform not handling " "all-NaN arrays as expected",
+    )
 
     # compute sigmas of transformed (normalized) values fitted to a gamma
     # distribution, using the full period of record as the calibration period
-    computed_values = compute.transform_fitted_gamma(precips_mm_monthly,
-                                                     data_year_start_monthly,
-                                                     data_year_start_monthly,
-                                                     data_year_end_monthly,
-                                                     compute.Periodicity.monthly)
-    np.testing.assert_allclose(computed_values,
-                               transformed_gamma_monthly,
-                               err_msg="Transformed gamma fitted monthly "
-                                       "values not computed as expected")
+    computed_values = compute.transform_fitted_gamma(
+        precips_mm_monthly,
+        data_year_start_monthly,
+        data_year_start_monthly,
+        data_year_end_monthly,
+        compute.Periodicity.monthly,
+    )
+    np.testing.assert_allclose(
+        computed_values,
+        transformed_gamma_monthly,
+        err_msg="Transformed gamma fitted monthly " "values not computed as expected",
+    )
 
     # compute sigmas of transformed (normalized) values fitted to a gamma
     # distribution, using the full period of record as the calibration period
-    computed_values = compute.transform_fitted_gamma(precips_mm_daily.flatten(),
-                                                     data_year_start_daily,
-                                                     calibration_year_start_daily,
-                                                     calibration_year_end_daily,
-                                                     compute.Periodicity.daily)
-    np.testing.assert_allclose(computed_values,
[…]
