+++ src/climate_indices/utils.py
@@ -8,9 +8,18 @@ import numpy as np
 
 # declare the function names that should be included in the public API for this module
 __all__ = [
-    "compute_days", "count_zeros_and_non_missings", "get_logger", "get_tolerance",
-    "gregorian_length_as_366day", "is_data_valid", "reshape_to_2d", "reshape_to_divs_years_months",
-    "rmse", "sign_change", "transform_to_366day", "transform_to_gregorian",
+    "compute_days",
+    "count_zeros_and_non_missings",
+    "get_logger",
+    "get_tolerance",
+    "gregorian_length_as_366day",
+    "is_data_valid",
+    "reshape_to_2d",
+    "reshape_to_divs_years_months",
+    "rmse",
+    "sign_change",
+    "transform_to_366day",
+    "transform_to_gregorian",
 ]
 
 
@@ -49,10 +58,7 @@ def compute_days(
 
     # loop over all time steps (months)
     for i in range(total_months):
-
-        years = int(
-            (i + initial_month - 1) / 12
-        )  # the number of years since the initial year
+        years = int((i + initial_month - 1) / 12)  # the number of years since the initial year
         months = int((i + initial_month - 1) % 12)  # the number of months since January
 
         # cook up a datetime object for the current time step (month)
@@ -157,12 +163,10 @@ def is_data_valid(
 
     # make sure we're not dealing with all NaN values
     if np.ma.isMaskedArray(data):
-
         # TODO fix this, therte is no ndarray.count according to PyCharm's warning, use another approach for this flag
         valid_flag = bool(data.count())
 
     elif isinstance(data, np.ndarray):
-
         valid_flag = not np.all(np.isnan(data))
 
     else:
@@ -205,8 +209,9 @@ def reshape_to_2d(
             # data is already in the shape we want, return it unaltered
             return values
         else:
-            message = "Values array has an invalid shape (2-D but second " + \
-                      f"dimension not {second_axis_length}): {shape}"
+            message = (
+                "Values array has an invalid shape (2-D but second " + f"dimension not {second_axis_length}): {shape}"
+            )
             _logger.error(message)
             raise ValueError(message)
 
@@ -221,10 +226,7 @@ def reshape_to_2d(
     final_year_values = shape[0] % second_axis_length
     if final_year_values > 0:
         pads = second_axis_length - final_year_values
-        values = np.pad(values,
-                        pad_width=(0, pads),
-                        mode="constant",
-                        constant_values=np.NaN)
+        values = np.pad(values, pad_width=(0, pads), mode="constant", constant_values=np.NaN)
 
     # we should have an ordinal number of years now
     # (ordinally divisible by second_axis_length)
@@ -254,8 +256,7 @@ def reshape_to_divs_years_months(
             # data is already in the shape we want, return it unaltered
             return monthly_values
         else:
-            message = "Values array has an invalid shape (3-D but " + \
-                      "third dimension is not 12): " + str(shape)
+            message = "Values array has an invalid shape (3-D but " + "third dimension is not 12): " + str(shape)
             _logger.error(message)
             raise ValueError(message)
 
@@ -268,8 +269,7 @@ def reshape_to_divs_years_months(
     # otherwise make sure that we've been passed in a 2-D array
     # of values with the final dimension size == 12
     elif shape[1] != 12:
-        message = "Values array has an invalid shape (second dimension " + \
-                  "should be 12, but is not): " + str(shape)
+        message = "Values array has an invalid shape (second dimension " + "should be 12, but is not): " + str(shape)
         _logger.error(message)
         raise ValueError(message)
 
@@ -285,12 +285,10 @@ def gregorian_length_as_366day(
     length_gregorian: int,
[…]
