+++ docs/conf.py
@@ -18,7 +18,7 @@ import sys
 import toml
 
 # since our source files are under a src folder we need to add that as our first path
-sys.path.insert(0, os.path.abspath('../src'))
+sys.path.insert(0, os.path.abspath("../src"))
 
 # -- Project information -----------------------------------------------------
 
@@ -29,10 +29,10 @@ author = "James Adams"
 pyproject = toml.load("../pyproject.toml")
 
 # get the full release version, including alpha/beta/rc tags
-release = pyproject['tool']['poetry']['version']
+release = pyproject["tool"]["poetry"]["version"]
 
 # get the major.minor version
-version = '.'.join(release.split('.')[:2])
+version = ".".join(release.split(".")[:2])
 
 # -- General configuration ---------------------------------------------------
 
@@ -149,9 +149,7 @@ latex_documents = [
 
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
-man_pages = [
-    (master_doc, "climate_indices", "climate_indices Documentation", [author], 1)
-]
+man_pages = [(master_doc, "climate_indices", "climate_indices Documentation", [author], 1)]
 
 
 # -- Options for Texinfo output ----------------------------------------------
