### 2023-09-19 12b7810
[522] black formatting, various other fixes suggested by PyCharm and SonarLint
