+++ tests/test_utils.py
@@ -64,28 +64,32 @@ def test_compute_days():
 def test_count_zeros_and_non_missings():
     # Test for the utils.count_zeros_and_non_missings() function
 
+    # messages used multiple times below
+    zero_count_error = "Failed to correctly count zero values"
+    non_missing_count_error = "Failed to correctly count non-missing values"
+    
     # vanilla use case
     values_list = [3, 4, 0, 2, 3.1, 5, np.NaN, 8, 5, 6, 0.0, np.NaN, 5.6, 2]
     values = np.array(values_list)
     zeros, non_missings = utils.count_zeros_and_non_missings(values)
     if zeros != 2:
-        raise AssertionError("Failed to correctly count zero values")
+        raise AssertionError(zero_count_error)
     if non_missings != 12:
-        raise AssertionError("Failed to correctly count non-missing values")
+        raise AssertionError(non_missing_count_error)
 
     # test with lists
     values = values_list
     zeros, non_missings = utils.count_zeros_and_non_missings(values)
     if zeros != 2:
-        raise AssertionError("Failed to correctly count zero values")
+        raise AssertionError(zero_count_error)
     if non_missings != 12:
-        raise AssertionError("Failed to correctly count non-missing values")
+        raise AssertionError(non_missing_count_error)
     values = [[3, 4, 0, 2, 3.1, 5, np.NaN], [8, 5, 6, 0.0, np.NaN, 5.6, 2]]
     zeros, non_missings = utils.count_zeros_and_non_missings(values)
     if zeros != 2:
-        raise AssertionError("Failed to correctly count zero values")
+        raise AssertionError(zero_count_error)
     if non_missings != 12:
-        raise AssertionError("Failed to correctly count non-missing values")
+        raise AssertionError(non_missing_count_error)
 
     # using a list that can't be converted
     # into an array should result in a TypeError
@@ -380,32 +384,32 @@ def test_rmse():
 def test_transform_to_gregorian():
     # Test for the utils.transform_to_gregorian() function
 
-    # an array of 366 values, representing a year with 366 days, such as a leap year
+    # an array of 366 values, representing a year with 366-days, such as a leap year
     values_366 = np.array(range(366))
 
     # an array of 365 values, representing a year with 365 days, with the value
-    # for all days after Feb 28th matching to those in the 366 day array
+    # for all days after Feb 28th matching to those in the 366-day array
     values_365 = np.array(range(365))
     values_365[59:] = [x + 1 for x in values_365[59:]]
 
-    # exercise the function with the 366 day year array,
+    # exercise the function with the 366-day year array,
     # using a non-leap year argument (1971)
     values_365_computed = utils.transform_to_gregorian(values_366, 1971)
 
     np.testing.assert_equal(
         values_365_computed,
         values_365,
-        "Not transforming the 1-D array of 366 days " + "into a corresponding 365 day array",
+        "Not transforming the 1-D array of 366-days " + "into a corresponding 365 day array",
     )
 
-    # exercise the function with the 366 day year array,
+    # exercise the function with the 366-day year array,
     # using a leap year argument (1972)
     values_366_computed = utils.transform_to_gregorian(values_366, 1972)
 
     np.testing.assert_equal(
         values_366_computed,
         values_366,
-        "Not transforming the 1-D array of " "366 days into a corresponding 366 day array",
+        "Not transforming the 1-D array of 366-days into a corresponding 366-day array",
     )
 
     # make sure that the function croaks with a ValueError
@@ -425,10 +429,10 @@ def test_transform_to_366day():
     # Test for the utils.transform_to_366day() function
 
[…]
