+++ tests/test_compute.py
@@ -5,12 +5,13 @@ import pytest
 
 from climate_indices import compute
 
-# ------------------------------------------------------------------------------
 # disable logging messages
 logging.disable(logging.CRITICAL)
 
+UNEXPECTED_PEARSON3_MESSAGE = "Transformed Pearson Type III fitted values not computed as expected"
+UNEXPECTED_SLIDING_SUMS_MESSAGE = "Sliding sums not computed as expected"
 
-# ------------------------------------------------------------------------------
+
 @pytest.mark.usefixtures(
     "precips_mm_monthly",
     "precips_mm_daily",
@@ -54,7 +55,7 @@ def test_transform_fitted_gamma(
         computed_values,
         all_nans,
         equal_nan=True,
-        err_msg="Gamma fit/transform not handling " "all-NaN arrays as expected",
+        err_msg="Gamma fit/transform not handling all-NaN arrays as expected",
     )
 
     # compute sigmas of transformed (normalized) values fitted to a gamma
@@ -69,7 +70,7 @@ def test_transform_fitted_gamma(
     np.testing.assert_allclose(
         computed_values,
         transformed_gamma_monthly,
-        err_msg="Transformed gamma fitted monthly " "values not computed as expected",
+        err_msg="Transformed gamma fitted monthly values not computed as expected",
     )
 
     # compute sigmas of transformed (normalized) values fitted to a gamma
@@ -86,7 +87,7 @@ def test_transform_fitted_gamma(
         transformed_gamma_daily,
         atol=0.001,
         equal_nan=True,
-        err_msg="Transformed gamma fitted daily " "values not computed as expected",
+        err_msg="Transformed gamma fitted daily values not computed as expected",
     )
 
     # confirm that we can call with a calibration period out of the valid range
@@ -103,7 +104,7 @@ def test_transform_fitted_gamma(
         transformed_gamma_monthly,
         atol=0.001,
         equal_nan=True,
-        err_msg="Transformed Pearson Type III " "fitted values not computed as expected",
+        err_msg=UNEXPECTED_PEARSON3_MESSAGE,
     )
 
     # if we provide a 1-D array then we need to provide a corresponding
@@ -140,7 +141,6 @@ def test_transform_fitted_gamma(
     )
 
 
-# ------------------------------------------------------------------------------
 @pytest.mark.usefixtures(
     "precips_mm_monthly",
     "precips_mm_daily",
@@ -216,7 +216,6 @@ def test_gamma_parameters(
     )
 
 
-# ------------------------------------------------------------------------------
 @pytest.mark.usefixtures(
     "precips_mm_monthly",
     "precips_mm_daily",
@@ -277,7 +276,7 @@ def test_transform_fitted_pearson(
         computed_values,
         expected_values,
         atol=0.001,
-        err_msg="Transformed Pearson Type III fitted " "values not computed as expected",
+        err_msg=UNEXPECTED_PEARSON3_MESSAGE,
     )
 
     # confirm that an input array of all NaNs will return the same array
@@ -293,7 +292,7 @@ def test_transform_fitted_pearson(
         computed_values,
         all_nans,
         equal_nan=True,
-        err_msg="Transformed Pearson Type III fitted " "values not computed as expected",
+        err_msg=UNEXPECTED_PEARSON3_MESSAGE,
     )
 
     # confirm that we can call with a calibration period outside of valid range
@@ -310,7 +309,7 @@ def test_transform_fitted_pearson(
         transformed_pearson3_monthly_fullperiod,
         atol=0.001,
         equal_nan=True,
-        err_msg="Transformed Pearson Type III fitted " "values not computed as expected",
+        err_msg=UNEXPECTED_PEARSON3_MESSAGE,
     )
 
[…]
