### 2023-11-05 05cd196
Update pyproject.toml
Recent xarray solved an issue that can affect the workflows based on climate indices
