+++ pyproject.toml
@@ -49,7 +49,7 @@ dask = '2022.2.0'
 h5netcdf = '1.1.0'
 python = '>=3.8,<3.12'
 scipy = '1.9.3'
-xarray = '2023.1.0'
+xarray = '2023.10.0'
 
 [tool.poetry.dev-dependencies]
 pytest = '*'
